python module/requirements.txt
requests
pandas
paho-mqtt